# specs/001-perf-problem-simulator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Performance Problem Simulator API
  description: |
    API for triggering and monitoring intentional performance problems in an Azure App Service
    application for educational and diagnostic demonstration purposes.
    
    ⚠️ **WARNING**: This API intentionally creates performance problems. It is designed for
    testing and learning purposes only. Do not use in production environments.
    
    ## Endpoint Categories
    
    - **Problem Triggers**: Endpoints that start performance problems
    - **Metrics**: Endpoints that report current application state
    - **Control**: Endpoints for managing simulations
    
    ## API Style
    
    This API uses action-oriented RPC-style endpoints (e.g., `/api/trigger-high-cpu`) rather than
    REST resource-based endpoints, as the operations are actions rather than resource management.
  version: 1.0.0
  contact:
    name: Performance Problem Simulator
  license:
    name: MIT
servers:
  - url: https://localhost:5001
    description: Local development
  - url: https://{appname}.azurewebsites.net
    description: Azure App Service
    variables:
      appname:
        default: perf-problem-simulator
        description: Azure App Service name
...
paths:
  /api/trigger-high-cpu:
    post:
      tags:
        - CPU
      summary: Trigger high CPU usage
      description: |
        Starts a CPU-intensive operation that spins all processor cores for the specified duration.
        The CPU usage will approach 100% and return to baseline when the duration expires.
        
        **Educational Note**: This endpoint uses `Parallel.For` with spin loops to saturate all
        available CPU cores, demonstrating how CPU-bound work affects application responsiveness.
      operationId: triggerHighCpu
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CpuStressRequest'
            examples:
              default:
                summary: Default 30-second CPU stress
                value:
                  durationSeconds: 30
              short:
                summary: Quick 10-second test
                value:
                  durationSeconds: 10
              extended:
                summary: Extended 5-minute stress
                value:
                  durationSeconds: 300
      responses:
        '200':
          description: CPU stress simulation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Problem endpoints are disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/reset-all:
    post:
      tags:
        - Control
      summary: Reset all simulations
      description: |
        Stops all active CPU and thread blocking simulations, releases all allocated memory,
        and optionally forces garbage collection. Use this to return the application to
        baseline state.
      operationId: resetAll
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                forceGarbageCollection:
                  type: boolean
                  default: true
      responses:
        '200':
          description: All simulations reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All simulations have been reset
                  simulationsStopped:
                    type: integer
                    example: 3
                  memoryReleasedBytes:
                    type: integer
                    format: int64
                    example: 524288000
...
components:
  schemas:
    CpuStressRequest:
      type: object
      properties:
        durationSeconds:
          type: integer
          minimum: 1
          maximum: 300
          default: 30
          description: Duration in seconds to sustain high CPU usage
...
    SimulationResult:
      type: object
      required:
        - simulationId
        - type
        - status
        - message
        - startedAt
      properties:
        simulationId:
          type: string
          format: uuid
          description: Unique identifier for this simulation
        type:
          type: string
          enum: [Cpu, Memory, ThreadBlock]
          description: Type of simulation
        status:
          type: string
          enum: [Started, Completed, Failed]
          description: Current status
        message:
          type: string
          description: Human-readable status message
        actualParameters:
          type: object
          additionalProperties: true
          description: Parameters actually used (may differ from request if values were capped)
        startedAt:
          type: string
          format: date-time
          description: When the simulation started
        estimatedEndAt:
          type: string
          format: date-time
          nullable: true
          description: When the simulation will complete (null for memory allocations)
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-level validation errors
        timestamp:
          type: string
          format: date-time